+++ client/components/NavBar/Forms/ViewForm.jsx
@@ -1,19 +1,19 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import NavBar from './NavBar';
+import NavBar from '../NavBar';
 
 
-const EditForm = (props) => {
+const ViewForm = (props) => {
 let { id } = useParams(); 
 console.log('props', props);
 console.log('topicId', id);
   return (
     <div>
       <NavBar />
-      <h1>Edit Form Made form testing purposes, sawpp </h1>
-      <p>{id}</p>
+      <h1>ViewForm Made form testing purposes, sup team </h1>
+      <h1>{id}</h1>
     </div>
   );
 }
 
-export default EditForm;
+export default ViewForm;
